+++ egon/map/layers.py
@@ -1,6 +1,8 @@
+
 import json
 from dataclasses import dataclass, field
 from itertools import product
+from enum import Enum
 from typing import List, Optional
 
 from django.db.models import IntegerField, BooleanField, Model, ObjectDoesNotExist
@@ -21,6 +23,13 @@ def get_opacity(source_layer):
     return LAYER_STYLES[source_layer]["paint"]["fill-opacity"]
 
 
+class VectorLayerType(Enum):
+    Fill = 0
+    Symbol = 1
+    Line = 2
+    Choropleth = 3
+
+
 @dataclass
 class VectorLayerData:
     source: str
@@ -30,6 +39,7 @@ class VectorLayerData:
     name_singular: str
     description: str
     clustered: bool = False
+    type: VectorLayerType = VectorLayerType.Fill
     popup_fields: list = field(default_factory=list)
 
 
@@ -47,6 +57,7 @@ class RasterLayerData:
 DEMAND: list = [
     VectorLayerData(
         source="demand_cts",
+        type=VectorLayerType.Choropleth,
         color=get_color("demand_cts"),
         model=models.DemandCts,
         name=_("Cts"),
@@ -61,6 +72,7 @@ DEMAND: list = [
     ),
     VectorLayerData(
         source="demand_household",
+        type=VectorLayerType.Choropleth,
         color="#fd8d3c",
         model=models.DemandHousehold,
         name=_("Haushalte"),
@@ -79,6 +91,7 @@ DEMAND: list = [
 GENERATION: list = [
     VectorLayerData(
         source="supply_biomass",
+        type=VectorLayerType.Symbol,
         color="blue",
         model=models.SupplyBiomass,
         name=_("Biomasse"),
@@ -89,6 +102,7 @@ GENERATION: list = [
     ),
     VectorLayerData(
         source="supply_run_of_river",
+        type=VectorLayerType.Symbol,
         color="blue",
         model=models.SupplyRunOfRiver,
         name=_("Hydro"),
@@ -98,6 +112,7 @@ GENERATION: list = [
     ),
     VectorLayerData(
         source="supply_wind",
+        type=VectorLayerType.Symbol,
         color="blue",
         model=models.SupplyWindOnshore,
         name=_("Wind onshore"),
@@ -107,6 +122,7 @@ GENERATION: list = [
     ),
     VectorLayerData(
         source="supply_solar",
+        type=VectorLayerType.Symbol,
         color="yellow",
         model=models.SupplySolarGround,
         name=_("Solar"),
@@ -144,6 +160,7 @@ GENERATION: list = [
 GRID: list = [
     VectorLayerData(
         source="ehv_line",
+        type=VectorLayerType.Line,
         color="blue",
         model=models.EHVLine,
         name=_("EHV Leitung"),
@@ -152,6 +169,7 @@ GRID: list = [
     ),
     VectorLayerData(
         source="hv_line",
+        type=VectorLayerType.Line,
         color="darkblue",
         model=models.HVLine,
         name=_("HV Leitung"),
@@ -160,6 +178,7 @@ GRID: list = [
     ),
     VectorLayerData(
         source="ehv_hv_station",
+        type=VectorLayerType.Symbol,
         color="green",
         model=models.EHVHVSubstation,
         name=_("EHV/HV Substation"),
@@ -168,6 +187,7 @@ GRID: list = [
     ),
     VectorLayerData(
         source="hv_mv_station",
+        type=VectorLayerType.Symbol,
         color="blue",
         model=models.HVMVSubstation,
         name=_("HV/MV Substation"),
@@ -179,6 +199,7 @@ GRID: list = [
 MODEL: list = [
     VectorLayerData(
         source="mv_grid_districts",
+        type=VectorLayerType.Choropleth,
         color="pink",
         model=models.MVGridDistricts,
         name=_("MV Grid Districts"),